Works on issue 165 by providing a warning if any lyrics are defined outside of lyric phrases before saving the chart.  Also, duplicated logic among the "Save" and "Save as" logic has been moved to eof_save_helper() for added consistency (ie. "Save as" hadn't been sorting/fixing lyrics).

+++ src/note.c
@@ -1143,7 +1143,7 @@ EOF_LYRIC_LINE *FindLyricLine_p(EOF_LYRIC * lp)
 	return NULL;	//No such line found
 }
 
-EOF_LYRIC_LINE *FindLyricLine(int lyricnum)
+EOF_LYRIC_LINE *FindLyricLine(unsigned long lyricnum)
 {
 	int linectr;
 	unsigned long lyricpos;
